Move bridgePortFromIfIndex and ifIndexFromBridgePort to BridgeMib trait
